Initialize sys_path_lock after the double fork
This change address a problem on MacOS.

If we create the asyncio lock before we do the double forks, it
actually initialize the asyncio context in the father process. And later
we get an "Bad file descriptor" when we try to start the asyncio endless
loop in the child process.

See: https://python-tornado.narkive.com/a9lCRPxp/errno-9-bad-file-descriptor#post2

Closes: #13