AnsibleTurboModule does transfer all AnsibleModule variable to the socket (#65)
AnsibleTurboModule does transfer all AnsibleModule variable to the socket

PASS_VARS is dict with each data item defined as a 2-tuple with first item corresponding to the attribute to the AnsibleModule class.
Consequently, only the check_mode was copied as this is the only variable with the same name on both the tuple and the PASS_VARS key.

Reviewed-by: Gonéri Le Bouder <goneri@lebouder.net>