Pass environment variables through task execution
Previously, the environment for a task would have been the same
environment as the first task run. This first task was responsible for
spawning the server which inherited that task's environment. This is a
problem in cases where the environment might change in a latter task,
for example, by setting the environment keyword on a task. In these
cases, the environment for the new task is lost.

This change wraps execution of a module in a temporary environment,
passing through whatever envvars might have been set by the controller.