Pass environment variables through task execution (#69)
Pass environment variables through task execution

Depends-On: ansible-collections/vmware.vmware_rest#235
SUMMARY

Previously, the environment for a task would have been the same
environment as the first task run. This first task was responsible for
spawning the server which inherited that task's environment. This is a
problem in cases where the environment might change in a latter task,
for example, by setting the environment keyword on a task. In these
cases, the environment for the new task is lost.
This change wraps execution of a module in a temporary environment,
passing through whatever envvars might have been set by the controller.

ISSUE TYPE


Bugfix Pull Request

COMPONENT NAME

ADDITIONAL INFORMATION

Reviewed-by: Gonéri Le Bouder <goneri@lebouder.net>
Reviewed-by: None <None>
Reviewed-by: Mike Graves <mgraves@redhat.com>