Update socket location calculation
Turbo modules rely on core Ansible mechanism for temporary path
creation on the target host. This means that the socket path
calculation must be in sync with the core calculation.

Up until now, turbo modules only handled one happy path: where remote
tmp location was set, it did not include any environment variables
that would need to be expaned, and it did not have a trailing slash.

This commit brings the core and turbo modules in sync when it comes to
remote temporary location calculations and adds a few tests to make
sure things work as expected.